pyproject.toml
[project]
name = "video-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "aiofiles>=25.1.0",
    "fastapi>=0.132.0",
    "open-clip-torch>=3.2.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "qwen-vl-utils>=0.0.14",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "uvicorn[standard]>=0.41.0",
]
